use size_t for 'offset'

+++ src/aabro.c
@@ -40,7 +40,7 @@
 
 abr_tree *abr_tree_malloc(
   short result,
-  int offset,
+  size_t offset,
   int length,
   char *note,
   abr_parser *p,
@@ -487,10 +487,10 @@ char *abr_parser_to_string(abr_parser *p)
 //
 // the parse methods
 
-typedef abr_tree *abr_p_func(char *, int, int, abr_parser *);
-abr_tree *abr_do_parse(char *input, int offset, int depth, abr_parser *p);
+typedef abr_tree *abr_p_func(char *, size_t, int, abr_parser *);
+abr_tree *abr_do_parse(char *input, size_t offset, int depth, abr_parser *p);
 
-abr_tree *abr_p_string(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_string(char *input, size_t offset, int depth, abr_parser *p)
 {
   char *s = p->string;
   int le = p->string_length;
@@ -506,7 +506,7 @@ abr_tree *abr_p_string(char *input, int offset, int depth, abr_parser *p)
   return abr_tree_malloc(su, offset, le, NULL, p, NULL);
 }
 
-abr_tree *abr_p_regex(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_regex(char *input, size_t offset, int depth, abr_parser *p)
 {
   regmatch_t ms[1];
 
@@ -520,11 +520,11 @@ abr_tree *abr_p_regex(char *input, int offset, int depth, abr_parser *p)
   return abr_tree_malloc(1, offset, ms[0].rm_eo - ms[0].rm_so, NULL, p, NULL);
 }
 
-abr_tree *abr_p_rep(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_rep(char *input, size_t offset, int depth, abr_parser *p)
 {
   int max = p->max;
   if (max < 0) max = MAX_REPS;
-  int off = offset;
+  size_t off = offset;
   size_t count = 0;
   int length = 0;
   abr_tree **reps = calloc(max + 1, sizeof(abr_tree *));
@@ -550,7 +550,7 @@ abr_tree *abr_p_rep(char *input, int offset, int depth, abr_parser *p)
   return abr_tree_malloc(result, offset, length, NULL, p, children);
 }
 
-abr_tree *abr_p_alt(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_alt(char *input, size_t offset, int depth, abr_parser *p)
 {
   size_t c = 0; while(1) if (p->children[c++] == NULL) break;
   abr_tree **ts = calloc(c, sizeof(abr_tree *));
@@ -571,14 +571,14 @@ abr_tree *abr_p_alt(char *input, int offset, int depth, abr_parser *p)
   return abr_tree_malloc(result, offset, length, NULL, p, ts);
 }
 
-abr_tree *abr_p_seq(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_seq(char *input, size_t offset, int depth, abr_parser *p)
 {
   size_t c = 0; while(1) if (p->children[c++] == NULL) break;
   abr_tree **ts = calloc(c, sizeof(abr_tree *));
 
   int result = 1;
   int length = 0;
-  int off = offset;
+  size_t off = offset;
 
   for (size_t i = 0; i < c - 1; i++)
   {
@@ -591,12 +591,12 @@ abr_tree *abr_p_seq(char *input, int offset, int depth, abr_parser *p)
   return abr_tree_malloc(result, offset, length, NULL, p, ts);
 }
 
-abr_tree *abr_p_not(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_not(char *input, size_t offset, int depth, abr_parser *p)
 {
   return NULL;
 }
 
-abr_tree *abr_p_name(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_name(char *input, size_t offset, int depth, abr_parser *p)
 {
   abr_tree *t = abr_do_parse(input, offset, depth + 1, p->children[0]);
   t->name = strdup(p->name);
@@ -604,17 +604,17 @@ abr_tree *abr_p_name(char *input, int offset, int depth, abr_parser *p)
   return t;
 }
 
-abr_tree *abr_p_presence(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_presence(char *input, size_t offset, int depth, abr_parser *p)
 {
   return NULL;
 }
 
-abr_tree *abr_p_absence(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_absence(char *input, size_t offset, int depth, abr_parser *p)
 {
   return NULL;
 }
 
-abr_tree *abr_p_n(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_p_n(char *input, size_t offset, int depth, abr_parser *p)
 {
   if (p->children[0] == NULL)
   {
@@ -639,7 +639,7 @@ abr_p_func *abr_p_funcs[] = { // const ?
   abr_p_n
 };
 
-abr_tree *abr_do_parse(char *input, int offset, int depth, abr_parser *p)
+abr_tree *abr_do_parse(char *input, size_t offset, int depth, abr_parser *p)
 {
   if (depth > MAX_DEPTH)
   {
@@ -650,12 +650,12 @@ abr_tree *abr_do_parse(char *input, int offset, int depth, abr_parser *p)
   return abr_p_funcs[p->type](input, offset, depth, p);
 }
 
-abr_tree *abr_parse(char *input, int offset, abr_parser *p)
+abr_tree *abr_parse(char *input, size_t offset, abr_parser *p)
 {
   return abr_do_parse(input, offset, 0, p);
 }
 
-abr_tree *abr_parse_all(char *input, int offset, abr_parser *p)
+abr_tree *abr_parse_all(char *input, size_t offset, abr_parser *p)
 {
   abr_tree *t = abr_do_parse(input, offset, 0, p);
